Messed with a bunch of things mostly category and browsing
deal with it

+++ WebContent/Category.jsp
@@ -4,7 +4,9 @@
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
+<%@include file="header.jsp" %>
 <%
+
 try {                 
                     Class.forName("org.postgresql.Driver");
 					Connection conn = null;
@@ -16,6 +18,7 @@ try {
             <%
             String action=request.getParameter("action");
             if (action != null && action.equals("insert")) {
+            	try{
             	System.out.println("gets into the intial part");
             	conn.setAutoCommit(false);
             	PreparedStatement pstmt =conn.prepareStatement(
@@ -25,22 +28,43 @@ try {
             	int rowCount= pstmt.executeUpdate();
             	conn.commit();
             	conn.setAutoCommit(true);
+            	}
+            	catch(SQLException e){
+            		%>name or description too long<%
+            	
+            	}
             }
             
             String deletion=request.getParameter("delete");
              if (deletion != null) {
+            	 try{
             	System.out.println("gets into the intial part delete "+request.getParameter("delete"));
             	conn.setAutoCommit(false);
+            	PreparedStatement ps =conn.prepareStatement(
+            			"SELECT * FROM products WHERE category=?");
+            	ps.setString(1,request.getParameter("delete").trim());
+            	ResultSet rq=ps.executeQuery();
+            	if(!rq.next()){          	
             	PreparedStatement pstmt =conn.prepareStatement(
             			"DELETE FROM category WHERE category.name= ?");
             	pstmt.setString(1,request.getParameter("delete").trim());
             	int rowCount= pstmt.executeUpdate();
+            
+            	}
+            	else{
+            		%>You can not delete <%=request.getParameter("delete").trim()%><%
+            	}
             	conn.commit();
             	conn.setAutoCommit(true);
+            	 }catch(SQLException e){
+            		 %>SQL delete problem <%
+            	
+            	 }
             }
             
              String update=request.getParameter("update");
              if (update != null) {
+            	 try{
             	System.out.println("gets into the intial part update"+request.getParameter("nameboxup"));
             	conn.setAutoCommit(false);
             	PreparedStatement pstmt =conn.prepareStatement(
@@ -50,9 +74,13 @@ try {
             	int rowCount= pstmt.executeUpdate();
             	conn.commit();
             	conn.setAutoCommit(true);
+            	 }
+            	 catch(SQLException e){
+            		 %>Description too long<%
+            	
+            	}
             }
-
-            
+             conn.setAutoCommit(true);
             %>
 <title>Insert title here</title>
 </head>
@@ -64,7 +92,8 @@ try {
 	<form method="POST" action="Category.jsp">
 	<input type="hidden" name="action" value="insert">
 	<input type="text" name="namebox">
-	<input type="text" name="dbox">
+	<textarea name="dbox" rows="1" cols="20"></textarea>
+	<!-- <input type="text" name="dbox"> -->
 	<button type="submit">Insert</button>	
 	</form>
 	<!--
@@ -87,6 +116,7 @@ try {
 		<th>description</th>
 	</tr>
 	<% 
+		conn.setAutoCommit(false);
 		ResultSet rs=null;
 		Statement statement=conn.createStatement();
 		rs=statement.executeQuery("SELECT * FROM category");
@@ -97,10 +127,20 @@ try {
 			<td><%=rs.getString("name") %></td>
 			<td><%=rs.getString("description") %></td>
 			<td>	
+			<%
+			PreparedStatement ps =conn.prepareStatement(
+            			"SELECT * FROM products WHERE category=?");
+            	ps.setString(1,rs.getString("name"));
+            	ResultSet rq=ps.executeQuery();
+            	if(!rq.next()){%>
 			<form method="POST" action="Category.jsp">
 			<input type="hidden" name="delete" value="<%=rs.getString("name")%>">
 			<button type="submit">Delete</button>
-			</form>		
+			</form>
+			<%}
+            	conn.setAutoCommit(true);
+			%>
+				
 			</td>
 			<td>
 				<form method="POST" action="Category.jsp">
